refactor: extract expandable-quote contract into dedicated module
Move EXPANDABLE_QUOTE_START/END sentinels and format_expandable_quote()
from providers/base.py into the new expandable_quote.py. Producers
(transcript_parser, codex, history) and consumers (entity_formatting)
now import from one place instead of through the provider boundary.
Also relocates UUID_RE from hook.py into providers/base.py where it
belongs, and removes the last raw-string imports from providers/__init__.py.
